+++ R/tbl_stack.R
@@ -191,31 +191,39 @@ tbl_stack <- function(tbls, group_header = NULL, quiet = FALSE, attr_order = seq
 
 # function prints changes to column labels and spanning headers
 .print_stack_differences <- function(tbls) {
-  tbl_differences <-
-    map2(
-      tbls, seq_len(length(tbls)),
-      ~ .x[["table_styling"]][["header"]] |>
-        dplyr::mutate(..tbl_id.. = .y)
+  any_header_difference <-
+    lapply(
+      tbls,
+      FUN = \(x) {
+        x[["table_styling"]][["header"]] |>
+          dplyr::filter(!hide) |>
+          dplyr::select("column", "label")
+      }
     ) |>
     dplyr::bind_rows() |>
-    dplyr::select("..tbl_id..", "column", "label") |>
-    tidyr::pivot_longer(cols = c("label")) |>
-    dplyr::group_by(.data$column, .data$name) |>
     dplyr::mutate(
-      new_value = .data$value[1],
-      name_fmt = dplyr::case_when(
-        name == "label" ~ "Column header"
-      )
+      .by = "column",
+      label_difference = .data$label != .data$label[1]
     ) |>
-    dplyr::filter(.data$new_value != .data$value) |>
-    dplyr::ungroup() |>
-    dplyr::arrange(.data$name != "label", .data$name_fmt, .data$..tbl_id..)
+    dplyr::pull("label_difference") |>
+    any()
 
-  if (nrow(tbl_differences) > 0) {
+  # if there are difference, print them to the console
+  if (any_header_difference) {
     cli::cli_inform(
-      c("Column headers among stacked tables differ. Headers from the first table are used.",
-        i = "Check the header is correct and use {.fun modify_header} to update,
-             or {.code quiet = TRUE} to suppress this message.")
+      c("Column headers among stacked tables differ.",
+        i = "Use {.fun modify_header} to update or {.code quiet = TRUE} to suppress this message.")
+    )
+
+    walk(
+      seq_along(tbls),
+      ~ tbls[[.x]] |>
+        getElement("table_styling") |>
+        getElement("header") |>
+        dplyr::filter(!hide) |>
+        dplyr::select("column", "label") |>
+        dplyr::mutate(label =  cli::cli_format(.data$label)) |>
+        tibble_as_cli(label = list(column = glue("Table {.x} Column Name"), label = "Header"))
     )
   }